+++ src/XrdNet/XrdNetLink.cc
@@ -265,7 +265,7 @@ int XrdNetLink::Send(const char *Buff, int Blen, int tmo)
 
    if (!Blen && !(Blen = strlen(Buff))) return 0;
    if ('\n' != Buff[Blen-1])
-      {const struct iovec iodata[2] = {{IOV_INIT((char *)Buff, Blen)},
+      {const struct iovec iodata[2] = {{IOV_INIT((char *)Buff, (size_t)Blen)},
                                        {IOV_INIT((char *)"\n", 1)}};
        return Send(iodata, 2, tmo);
       }
@@ -315,7 +315,7 @@ int XrdNetLink::Send(const char *dest, const char *Buff, int Blen, int tmo)
 
    if (!Blen && !(Blen = strlen(Buff))) return 0;
    if ('\n' != Buff[Blen-1])
-      {const struct iovec iodata[2] = {{IOV_INIT((char *)Buff, Blen)},
+      {const struct iovec iodata[2] = {{IOV_INIT((char *)Buff, (size_t)Blen)},
                                        {IOV_INIT((char *)"\n", 1)}};
        return Send(dest, iodata, 2, tmo);
       }
